+++ src/main/java/com/pax/market/api/sdk/java/api/emm/emmApp/dto/EmmAppAvailableTestVersionResponse.java
@@ -1,0 +1,8 @@
+package com.pax.market.api.sdk.java.api.emm.emmApp.dto;
+
+import com.pax.market.api.sdk.java.api.base.dto.Response;
+
+public class EmmAppAvailableTestVersionResponse extends Response<EmmAppAvailableTestVersionDTO> {
+
+    private static final long serialVersionUID = 8527439089855890429L;
+}