+++ src/main/java/com/pax/market/api/sdk/java/api/terminalApk/TerminalApkApi.java
@@ -14,6 +14,7 @@ package com.pax.market.api.sdk.java.api.terminalApk;
 
 import com.google.gson.Gson;
 import com.pax.market.api.sdk.java.api.BaseThirdPartySysApi;
+import com.pax.market.api.sdk.java.api.base.dto.EmptyResponse;
 import com.pax.market.api.sdk.java.api.base.dto.PageRequestDTO;
 import com.pax.market.api.sdk.java.api.base.dto.Response;
 import com.pax.market.api.sdk.java.api.base.dto.Result;
@@ -29,8 +30,11 @@ import com.pax.market.api.sdk.java.api.util.StringUtils;
 import com.pax.market.api.sdk.java.api.validate.Validators;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
+import java.util.Objects;
 
 /**
  * Description
@@ -46,6 +50,7 @@ public class TerminalApkApi extends BaseThirdPartySysApi{
 	private static final String CREATE_TERMINAL_APK_URL = "/v1/3rdsys/terminalApks";
 	private static final String GET_TERMINAL_APK_URL = "/v1/3rdsys/terminalApks/{terminalApkId}";
 	private static final String SUSPEND_TERMINAL_APK_URL = "/v1/3rdsys/terminalApks/suspend";
+    private static final String DELETE_TERMINAL_APK_URL = "/v1/3rdsys/terminalApks/{terminalApkId}";
 	private static final String UNINSTALL_TERMINAL_APK_URL = "/v1/3rdsys/terminalApks/uninstall";
 
 
@@ -57,8 +62,17 @@ public class TerminalApkApi extends BaseThirdPartySysApi{
 		super(baseUrl, apiKey, apiSecret, locale);
 	}
 
-	public Result<TerminalApkDTO> searchTerminalApk(int pageNo, int pageSize, SearchOrderBy orderBy,
+    public Result<TerminalApkDTO> searchTerminalApk(int pageNo, int pageSize, SearchOrderBy orderBy,
                                                     String terminalTid, String appPackageName, PushStatus status){
+        return searchTerminalApk(pageNo, pageSize, orderBy, terminalTid, appPackageName, status, null,null);
+    }
+
+    public Result<TerminalApkDTO> searchTerminalApk(int pageNo, int pageSize, SearchOrderBy orderBy,
+                                                    String terminalTid, String appPackageName, PushStatus status, List<String> pidList){
+        return searchTerminalApk(pageNo, pageSize, orderBy, terminalTid, appPackageName, status, null, pidList);
+    }
+	public Result<TerminalApkDTO> searchTerminalApk(int pageNo, int pageSize, SearchOrderBy orderBy,
+                                                    String terminalTid, String appPackageName, PushStatus status, String serialNo, List<String> pidList){
         ThirdPartySysApiClient client = new ThirdPartySysApiClient(getBaseUrl(), getApiKey(), getApiSecret());
         PageRequestDTO page = new PageRequestDTO();
         page.setPageNo(pageNo);
@@ -71,14 +85,18 @@ public class TerminalApkApi extends BaseThirdPartySysApi{
         if(!validationErrs.isEmpty()) {
             return new Result<>(validationErrs);
         }
-        if(StringUtils.isEmpty(terminalTid)) {
-            validationErrs.add(getMessage("parameter.not.null", "terminalTid"));
+        if(StringUtils.isEmpty(terminalTid) && StringUtils.isEmpty(serialNo)) {
+            validationErrs.add(getMessage("parameter.sn.tid.empty"));
             return new Result<>(validationErrs);
         }
 
         SdkRequest request = getPageRequest(SEARCH_TERMINAL_APK_LIST_URL, page);
         request.addRequestParam("terminalTid", terminalTid);
+        request.addRequestParam("serialNo", serialNo);
         request.addRequestParam("appPackageName", appPackageName);
+        if (Objects.nonNull(pidList) && !pidList.isEmpty()){
+            request.addRequestParam("pidList",  StringUtils.join(pidList, ","));
+        }
         if(status != null){
             request.addRequestParam("status", status.val());
         }
@@ -147,6 +165,19 @@ public class TerminalApkApi extends BaseThirdPartySysApi{
         request.setRequestBody(new Gson().toJson(uninstallApkRequest, UpdateTerminalApkRequest.class));
         Response resp = EnhancedJsonUtils.fromJson(client.execute(request), Response.class);
         return new Result<>(resp);
+    }
+
+    public Result<String> deleteTerminalApk(Long terminalApkId) {
+        logger.debug("terminalApkId= {}", terminalApkId);
+        List<String> validationErrs = Validators.validateId(terminalApkId, "parameter.id.invalid", "terminalApkId");
+        if(!validationErrs.isEmpty()) {
+            return new Result<>(validationErrs);
+        }
+        ThirdPartySysApiClient client = new ThirdPartySysApiClient(getBaseUrl(), getApiKey(), getApiSecret());
+        SdkRequest request = createSdkRequest(DELETE_TERMINAL_APK_URL.replace("{terminalApkId}", terminalApkId.toString()));
+        request.setRequestMethod(SdkRequest.RequestMethod.DELETE);
+        EmptyResponse emptyResponse = EnhancedJsonUtils.fromJson(client.execute(request), EmptyResponse.class);
+        return new Result<>(emptyResponse);
     }
 
     public enum SearchOrderBy {