+++ src/main/java/com/pax/market/api/sdk/java/api/terminalApkParameter/dto/UpdateApkParameterRequest.java
@@ -37,6 +37,7 @@ public class UpdateApkParameterRequest  implements Serializable {
 
     private List<FileParameter> base64FileParameters;
 
+    private Boolean validateUndefinedParameter;
 
 
     public String getParamTemplateName() {
@@ -61,5 +62,13 @@ public class UpdateApkParameterRequest  implements Serializable {
 
     public void setBase64FileParameters(List<FileParameter> base64FileParameters) {
         this.base64FileParameters = base64FileParameters;
+    }
+
+    public Boolean getValidateUndefinedParameter() {
+        return validateUndefinedParameter;
+    }
+
+    public void setValidateUndefinedParameter(Boolean validateUndefinedParameter) {
+        this.validateUndefinedParameter = validateUndefinedParameter;
     }
 }